Move machine-specific files to test/machine.ts
